rust: add "optimized custom hashmap" program
This program is an approximate port from the optimized C program.
It is actually a bit slower, but if you compile the C program with
Clang, then performance is about on par. I didn't dig in any
further, but gcc might just have better codegen in this case than
LLVM does.