rust: add "bonus" program
This program intentional flouts the constraints laid down in the blog
post, and instead proposes a new challenge itself: write a program that
is fully Unicode aware. My hypothesis is that porting this to other
languages may be quite a bit more difficult.

Note that in the test script, we compile the program but do not check
its output. That's because its output is intentionally different than
the other programs.