Fix (and simplify!) optimized Go version per Miguel's version
This fixes a subtle bug in mine that resulted in incorrect output
*some* of the time -- when a partial Read() returned a small amount
of data with no LF character in it. If `Read` did a partial read that
didn't include a linefeed character, we'd split a word in the middle
and count it as two words. Partial reads aren't common (and never
happened in my testing), but definitely do happen. Fun!